requirements.txt

---

### **requirements.txt**
```plaintext
opencv-python
pillow
torch
torchvision
diffusers
transformers
rembg
numpy

